Small dynamic change
Image of the highlighted god from the dropdown will now pop up for the user to see more clearly as well as the name appears in the selected option part. If the user moves somewhere else the original selected god will be put back. If a new god is selected, everything is updated